Add product search request

+++ app/src/main/java/com/example/hikeculator/domain/entities/Convertor.kt
@@ -2,7 +2,7 @@ package com.example.hikeculator.domain.entities
 
 class Convertor {
 
-    fun getTotalColorizes(product: Product): Long = product.nutritionalValue.calories
+    fun getTotalColorizes(product: Product): Double = product.nutritionalValue.calories
 
-    fun getTotalColorizes(dayMeal: DayMeal): Long = dayMeal.totalCalories
+    fun getTotalColorizes(dayMeal: DayMeal): Double = dayMeal.totalCalories
 }
